% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiktive-package.R
\docType{package}
\name{fiktive-package}
\alias{fiktive}
\alias{fiktive-package}
\title{fiktive: Fictitious Danish Register Data}
\description{
Generate structurally valid fictitious Danish register data, for writing
and checking analysis code outside Statistics Denmark. This package
creates data; it does not extract from real registers.
}
\author{
\strong{Maintainer}: Sara Schwartz \email{saras@clin.au.dk}

}
\keyword{internal}
