\name{scenario_independence}
\alias{scenario_independence}
\title{Independence scenario object}
\usage{
scenario_independence(version = 1L)
}
\arguments{
\item{version}{Integer scenario version (default \code{1L}).}
}
\value{
A list of class \code{fiktive_scenario}.
}
\description{
\code{scenario = NULL} on generators is independence. Associations /
confounders / biases stay empty until later steps.
}
